Proyecto finalizado API 28 -Falta concretar en otras API's y mejorar rendimiento de imagenes.

+++ app/src/main/java/com/example/foodtrick/Objetos/productoMostrar.java
@@ -5,14 +5,14 @@ public class productoMostrar {
     private String nombre;
     private String categoria;
     private int img;
-    private int grasas;
-    private int hidratos;
-    private int azucares;
+    private float grasas;
+    private float hidratos;
+    private float azucares;
 
     public productoMostrar() {
     }
 
-    public productoMostrar(String nombre, String categoria, int img, int grasas, int hidratos, int azucares) {
+    public productoMostrar(String nombre, String categoria, int img, float grasas, float hidratos, float azucares) {
         this.nombre = nombre;
         this.categoria = categoria;
         this.img = img;
@@ -45,27 +45,27 @@ public class productoMostrar {
         this.img = img;
     }
 
-    public int getGrasas() {
+    public float getGrasas() {
         return grasas;
     }
 
-    public void setGrasas(int grasas) {
+    public void setGrasas(float grasas) {
         this.grasas = grasas;
     }
 
-    public int getHidratos() {
+    public float getHidratos() {
         return hidratos;
     }
 
-    public void setHidratos(int hidratos) {
+    public void setHidratos(float hidratos) {
         this.hidratos = hidratos;
     }
 
-    public int getAzucares() {
+    public float getAzucares() {
         return azucares;
     }
 
-    public void setAzucares(int azucares) {
+    public void setAzucares(float azucares) {
         this.azucares = azucares;
     }
 }
